Create README.md
Welcome to the Disease Detection and Medicine Recommendation System project! In this project, we aim to develop an intelligent system that detects diseases based on provided Symptoms of the patient and recommends suitable medications. This project integrates advanced machine learning techniques to enhance healthcare by automating the diagnosis process and suggesting personalized medical solutions. Get ready to explore the world of data-driven healthcare solutions and a journey to create a system that contributes to better patient care